+++ zict/tests/test_file.py
@@ -1,5 +1,5 @@
 import os
-import shutil
+import pathlib
 
 import pytest
 
@@ -7,33 +7,22 @@ from zict.file import File
 from zict.tests import utils_test
 
 
-@pytest.fixture
-def fn():
-    filename = ".tmp"
-    if os.path.exists(filename):
-        shutil.rmtree(filename)
-
-    yield filename
-
-    if os.path.exists(filename):
-        shutil.rmtree(filename)
-
-
-def test_mapping(fn):
+def test_mapping(tmpdir):
     """
     Test mapping interface for File().
     """
-    z = File(fn)
+    z = File(tmpdir)
     utils_test.check_mapping(z)
 
 
-def test_implementation(fn):
-    z = File(fn)
+@pytest.mark.parametrize("dirtype", [str, pathlib.Path, lambda x: x])
+def test_implementation(tmpdir, dirtype):
+    z = File(dirtype(tmpdir))
     assert not z
 
     z["x"] = b"123"
-    assert os.listdir(fn) == ["x"]
-    with open(os.path.join(fn, "x"), "rb") as f:
+    assert os.listdir(tmpdir) == ["x"]
+    with open(tmpdir / "x", "rb") as f:
         assert f.read() == b"123"
 
     assert "x" in z
@@ -42,14 +31,14 @@ def test_implementation(fn):
     assert out == b"123"
 
 
-def test_memmap_implementation(fn):
-    z = File(fn, memmap=True)
+def test_memmap_implementation(tmpdir):
+    z = File(tmpdir, memmap=True)
     assert not z
 
     mv = memoryview(b"123")
     assert "x" not in z
     z["x"] = mv
-    assert os.listdir(fn) == ["x"]
+    assert os.listdir(tmpdir) == ["x"]
     assert "x" in z
     mv2 = z["x"]
     assert mv2 == b"123"
@@ -58,43 +47,43 @@ def test_memmap_implementation(fn):
     assert mv2 == b"223"
 
 
-def test_str(fn):
-    z = File(fn)
-    assert str(z) == repr(z) == f"<File: {fn}, 0 elements>"
+def test_str(tmpdir):
+    z = File(tmpdir)
+    assert str(z) == repr(z) == f"<File: {tmpdir}, 0 elements>"
 
 
-def test_setitem_typeerror(fn):
-    z = File(fn)
+def test_setitem_typeerror(tmpdir):
+    z = File(tmpdir)
     with pytest.raises(TypeError):
         z["x"] = 123
 
 
-def test_contextmanager(fn):
-    with File(fn) as z:
+def test_contextmanager(tmpdir):
+    with File(tmpdir) as z:
         z["x"] = b"123"
 
-    with open(os.path.join(fn, "x"), "rb") as f:
-        assert f.read() == b"123"
+    with open(tmpdir / "x", "rb") as fh:
+        assert fh.read() == b"123"
 
 
-def test_delitem(fn):
-    z = File(fn)
+def test_delitem(tmpdir):
+    z = File(tmpdir)
 
     z["x"] = b"123"
-    assert os.path.exists(os.path.join(z.directory, "x"))
+    assert os.listdir(tmpdir) == ["x"]
     del z["x"]
-    assert not os.path.exists(os.path.join(z.directory, "x"))
+    assert os.listdir(tmpdir) == []
 
 
-def test_missing_key(fn):
-    z = File(fn)
+def test_missing_key(tmpdir):
+    z = File(tmpdir)
 
     with pytest.raises(KeyError):
         z["x"]
 
 
-def test_arbitrary_chars(fn):
-    z = File(fn)
+def test_arbitrary_chars(tmpdir):
+    z = File(tmpdir)
 
     # Avoid hitting the Windows max filename length
     chunk = 16
@@ -109,7 +98,7 @@ def test_arbitrary_chars(fn):
         assert list(z.items()) == [(key, b"foo")]
         assert list(z.values()) == [b"foo"]
 
-        zz = File(fn)
+        zz = File(tmpdir)
         assert zz[key] == b"foo"
         assert list(zz) == [key]
         assert list(zz.keys()) == [key]
@@ -122,8 +111,8 @@ def test_arbitrary_chars(fn):
             z[key]
 
 
-def test_write_list_of_bytes(fn):
-    z = File(fn)
+def test_write_list_of_bytes(tmpdir):
+    z = File(tmpdir)
 
     z["x"] = [b"123", b"4567"]
     assert z["x"] == b"1234567"